added sfu

+++ main.py
@@ -3,11 +3,11 @@ from flask_socketio import SocketIO, emit, send
 import os
 
 database = []
+sfulist= []
 
 con = Flask(__name__)
 con.secretKey = os.environ.get('Secret_Key')
 socket = SocketIO(con)
-
 
 @con.errorhandler(404)
 def errpage(e):
@@ -18,8 +18,13 @@ def about():
     return render_template('about.html')
 
 @con.route('/')
+@con.route('/home')
 def home():
     return render_template('home.html')
+
+@con.route('/sfu')
+def sfupage():
+    return render_template('sfu.html')
 
 @socket.on_error_default
 def default_error_handler(e):
@@ -48,6 +53,11 @@ def Credentials(credential_json):
                     break
         else:
             emit('flashing', {'messageid':3, 'message':'No such concurrent Meeting'})
+
+@socket.on('whatsmysid')
+def sid():
+    sid = request.sid
+    emit('yoursidis', {'sid': sid}, room = sid)
 
 @socket.on('declined')
 def declined(message):
@@ -104,13 +114,17 @@ def special(message):
 
 @socket.on('disconnect')
 def close():
+    global sfulist
     rem = request.sid
     emit('close', rem) # to room needed.
     print('Disconnection from', rem)
-    for i,j in enumerate(database):
-        if j['sessionid'] == rem:
-            database.remove(database[i])
-            break
+    if rem in sfulist:
+        sfulist = []
+    else:
+        for i,j in enumerate(database):
+            if j['sessionid'] == rem:
+                database.remove(database[i])
+                break
 
 @socket.on('hangup')
 def hangup(to):
@@ -118,6 +132,27 @@ def hangup(to):
     for i in to:
         emit('hangupreceived', sid, room = i)
 
+@socket.on('readysfu') # from SFU page.
+def sfus():
+    print('SFUid ', request.sid)
+    sfulist.append(request.sid)
+
+@socket.on('joinsfu') 
+def sfu(message):
+    for i in message['to']:
+        print('Sending query to Join SFU to ', i)
+        emit('pleasejoinsfu',{'sfu': sfulist[0]}, room = i)# to members
+
+@socket.on('offeringtosfu')
+def offersfu(message):
+    sid = request.sid
+    print('Offering to sfu by',sid)
+    for i in database:
+        if i['sessionid'] == sid:
+            clientname = i['username']
+            break
+    emit('offertojoinsfu', {'callerid':sid, 'message': message['message'], 'name': clientname}, room = sfulist[0])#to SFU
+
 if __name__ == '__main__':
-    socket.run(con, debug = True) # for locally , host='192.168.1.9'
+    socket.run(con, debug = True, host='192.168.1.9') # for locally , host='192.168.1.9'
     #con.run() # for heroku
